+++ ios/Classes/PNSBuildModelUtils.m
@@ -285,7 +285,9 @@
   Boolean isDialog = [viewConfig boolValueForKey: @"isDialog" defaultValue: NO];
   /// 导航设置
   model.navIsHidden = [viewConfig boolValueForKey: @"navHidden" defaultValue: NO];
-  model.navColor = [self colorWithHexString: [viewConfig stringValueForKey: @"navColor" defaultValue: @"0x3971fe"] alpha: 1];
+  model.navColor = [self getColor: [viewConfig stringValueForKey: @"navColor" defaultValue: @"0x3971fe"]];
+
+//  [self colorWithHexString: [viewConfig stringValueForKey: @"navColor" defaultValue: @"0x3971fe"] alpha: 1];
   model.navTitle = [
     [NSAttributedString alloc]
       initWithString: [viewConfig stringValueForKey: @"navText" defaultValue: @"一键登录"]
@@ -296,11 +298,31 @@
   ];
   
   /// 返回按钮
-  model.hideNavBackItem = [viewConfig boolValueForKey: @"navReturnHidden" defaultValue: NO];
+  bool isHiddenNavBack = [viewConfig boolValueForKey: @"navReturnHidden" defaultValue: NO];
+  bool isCustomNavBack = [viewConfig boolValueForKey: @"customPageBackgroundLyout" defaultValue: NO];
+  model.hideNavBackItem = isHiddenNavBack;
   /// 动态读取assets文件夹下的资源
   UIImage * navBackImage = [self changeUriPathToImage: viewConfig[@"navReturnImgPath"]];
   if(navBackImage != nil){
     model.navBackImage = navBackImage;
+  }
+  if (isCustomNavBack) {
+    /// 自定义返回按钮
+    model.navBackButtonFrameBlock = ^CGRect(CGSize screenSize, CGSize superViewSize, CGRect frame) {
+      UIImageView *imageView = [[UIImageView alloc]init];
+      imageView.image = navBackImage;
+      imageView.frame = CGRectMake(CGRectGetMinX(frame),
+                                         CGRectGetMaxY(frame),
+                                         CGRectGetWidth(frame),
+                                   CGRectGetHeight(frame));
+      
+      frame.origin.y = [viewConfig floatValueForKey: @"navReturnOffsetY" defaultValue: 5];
+      frame.origin.x = [viewConfig floatValueForKey: @"navReturnOffsetX" defaultValue: 15];
+      
+      frame.size.width = [viewConfig floatValueForKey: @"navReturnImgWidth" defaultValue: 40];
+      frame.size.height = [viewConfig floatValueForKey: @"navReturnImgHeight" defaultValue: 40];
+      return frame;
+    };
   }
   
   /// 右侧按钮布局设置
@@ -309,13 +331,13 @@
   // model.navMoreView = rightBtn;
   
   /// 协议页面导航设置
-  model.privacyNavColor = [self colorWithHexString: [viewConfig stringValueForKey: @"webNavColor" defaultValue: @"#000"] alpha: 1];
+  model.privacyNavColor = [self getColor: [viewConfig stringValueForKey: @"webNavColor" defaultValue: @"#000"]];
   UIImage * privacyNavBackImage = [self changeUriPathToImage: viewConfig[@"webNavReturnImgPath"]];
   if(privacyNavBackImage != nil){
     model.privacyNavBackImage = privacyNavBackImage;
   }
   model.privacyNavTitleFont = [UIFont systemFontOfSize: [viewConfig floatValueForKey: @"webNavTextSize" defaultValue: 18]];
-  model.privacyNavTitleColor = [self colorWithHexString: [viewConfig stringValueForKey: @"webNavTextColor" defaultValue: @"#000"] alpha: 1];
+  model.privacyNavTitleColor = [self getColor: [viewConfig stringValueForKey: @"webNavTextColor" defaultValue: @"#000"]];
   
   /// logo 设置
   model.logoIsHidden = [viewConfig boolValueForKey: @"logoHidden" defaultValue: NO];
@@ -354,7 +376,8 @@
   };
   
   /// number 设置
-  model.numberColor = [self colorWithHexString: [viewConfig stringValueForKey: @"numberColor" defaultValue: @"555"] alpha: 1];
+  model.numberColor = [self getColor: [viewConfig stringValueForKey: @"numberColor" defaultValue: @"#555"]];
+  
   model.numberFont = [UIFont systemFontOfSize: [viewConfig floatValueForKey: @"numberSize" defaultValue: 17]];
   model.numberFrameBlock = ^CGRect(CGSize screenSize, CGSize superViewSize, CGRect frame) {
       if (screenSize.height > screenSize.width) {
@@ -370,7 +393,7 @@
     [NSAttributedString alloc]
         initWithString: [viewConfig stringValueForKey: @"logBtnText" defaultValue: @"一键登录欢迎语"]
             attributes: @{
-              NSForegroundColorAttributeName: [self colorWithHexString: [viewConfig stringValueForKey: @"logBtnTextColor" defaultValue: @"ff00ff"] alpha: 1],
+              NSForegroundColorAttributeName: [self getColor: [viewConfig stringValueForKey: @"logBtnTextColor" defaultValue: @"#ff00ff"]],
               NSFontAttributeName: [UIFont systemFontOfSize: [viewConfig floatValueForKey: @"logBtnTextSize" defaultValue: 23]]
             }
   ];
@@ -389,14 +412,14 @@
     
     UIImage *defaultClick = [UIImage imageNamed:@"button_click"];
     
-    UIImage *defaultUnClick = [UIImage imageNamed:@"button_click"];
+    UIImage *defaultUnClick = [UIImage imageNamed:@"button_unclick"];
     
     if ((login_btn_normal != nil || defaultClick != nil) && (login_btn_unable != nil || defaultUnClick != nil) && (login_btn_press != nil || defaultClick != nil)) {
       // 登录按钮设置
       model.loginBtnBgImgs = @[
-        login_btn_normal?:[UIImage imageNamed:@"button_click"],
-        login_btn_unable?:[UIImage imageNamed:@"button_unclick"],
-        login_btn_press?:[UIImage imageNamed:@"button_click"]
+        login_btn_normal?:defaultClick,
+        login_btn_unable?:defaultUnClick,
+        login_btn_press?:defaultClick
       ];
     }
   }
@@ -440,10 +463,9 @@
   BOOL checkStatus = [viewConfig boolValueForKey: @"checkBoxHidden" defaultValue: NO];
   model.checkBoxIsHidden = checkStatus;
   if (!checkStatus) {
-    NSArray *checkImagePath = [[viewConfig stringValueForKey: @"checkBoxImages" defaultValue: @","] componentsSeparatedByString:@","];
-    if (checkImagePath.count == 2) {
-      UIImage* unchecked = [self changeUriPathToImage: checkImagePath[0]];
-      UIImage* checked = [self changeUriPathToImage: checkImagePath[1]];
+    UIImage* unchecked = [self changeUriPathToImage: [viewConfig stringValueForKey: @"uncheckedImgPath" defaultValue: nil]];
+    UIImage* checked = [self changeUriPathToImage: [viewConfig stringValueForKey: @"checkedImgPath" defaultValue: nil]];
+    if (unchecked != nil && checked != nil) {
       model.checkBoxImages = @[
         unchecked,
         checked
@@ -602,7 +624,7 @@
     TXCustomModel *model = [[TXCustomModel alloc] init];
     
   model.alertBarIsHidden = [viewConfig boolValueForKey: @"alertBarIsHidden" defaultValue: NO];
-  model.alertTitleBarColor = [self colorWithHexString: [viewConfig stringValueForKey: @"alertTitleBarColor" defaultValue: @"0x3971fe"] alpha: 1];
+  model.alertTitleBarColor = [self getColor: [viewConfig stringValueForKey: @"alertTitleBarColor" defaultValue: @"0x3971fe"]];
   model.alertTitle = [
     [NSAttributedString alloc]
       initWithString: [viewConfig stringValueForKey: @"navText" defaultValue: @"一键登录"]
@@ -616,20 +638,20 @@
   UIImage * alertCloseImage = [self changeUriPathToImage: viewConfig[@"alertCloseImage"]];
   model.alertCloseImage = alertCloseImage?:[UIImage imageNamed:@"icon_close_light"];
   
-  model.alertBlurViewColor = [self colorWithHexString: [viewConfig stringValueForKey: @"alertBlurViewColor" defaultValue: @"#000000"] alpha: 1];
+  model.alertBlurViewColor = [self getColor: [viewConfig stringValueForKey: @"alertBlurViewColor" defaultValue: @"#000000"]];
   model.alertBlurViewAlpha = [viewConfig floatValueForKey: @"alertBlurViewAlpha" defaultValue: 0.5];
   NSString *radiuString = [viewConfig stringValueForKey: @"alertCornerRadiusArray" defaultValue: @"10,10,10,10"];
   NSArray *alertCornerRadiusArray = [radiuString componentsSeparatedByString: @","];
   model.alertCornerRadiusArray = [self _map: alertCornerRadiusArray]; //@[@10, @10, @10, @10];
     
   /// 协议页面导航设置
-  model.privacyNavColor = [self colorWithHexString: [viewConfig stringValueForKey: @"webNavColor" defaultValue: @"#000"] alpha: 1];
+  model.privacyNavColor =  [self getColor: [viewConfig stringValueForKey: @"webNavColor" defaultValue: @"#000000"]];
   UIImage * privacyNavBackImage = [self changeUriPathToImage: viewConfig[@"webNavReturnImgPath"]];
   if(privacyNavBackImage != nil){
     model.privacyNavBackImage = privacyNavBackImage;
   }
   model.privacyNavTitleFont = [UIFont systemFontOfSize: [viewConfig floatValueForKey: @"webNavTextSize" defaultValue: 18]];
-  model.privacyNavTitleColor = [self colorWithHexString: [viewConfig stringValueForKey: @"webNavTextColor" defaultValue: @"#000"] alpha: 1];
+  model.privacyNavTitleColor = [self getColor: [viewConfig stringValueForKey: @"webNavTextColor" defaultValue: @"#000000"]];
   
   /// logo 设置
   model.logoIsHidden = [viewConfig boolValueForKey: @"logoHidden" defaultValue: NO];
@@ -669,7 +691,7 @@
   };
   
   /// number 设置
-  model.numberColor = [self colorWithHexString: [viewConfig stringValueForKey: @"numberColor" defaultValue: @"555"] alpha: 1];
+  model.numberColor = [self getColor: [viewConfig stringValueForKey: @"numberColor" defaultValue: @"#555"]];
   model.numberFont = [UIFont systemFontOfSize: [viewConfig floatValueForKey: @"numberSize" defaultValue: 17]];
   model.numberFrameBlock = ^CGRect(CGSize screenSize, CGSize superViewSize, CGRect frame) {
       if (screenSize.height > screenSize.width) {
@@ -685,7 +707,7 @@
     [NSAttributedString alloc]
         initWithString: [viewConfig stringValueForKey: @"logBtnText" defaultValue: @"一键登录欢迎语"]
             attributes: @{
-              NSForegroundColorAttributeName: [self colorWithHexString: [viewConfig stringValueForKey: @"logBtnTextColor" defaultValue: @"ff00ff"] alpha: 1],
+              NSForegroundColorAttributeName: [self getColor: [viewConfig stringValueForKey: @"logBtnTextColor" defaultValue: @"#ff00ff"]],
               NSFontAttributeName: [UIFont systemFontOfSize: [viewConfig floatValueForKey: @"logBtnTextSize" defaultValue: 23]]
             }
   ];
@@ -703,16 +725,16 @@
     UIImage * login_btn_press = [self changeUriPathToImage: logBtnCustomBackgroundImagePath[2]];
     
     // default
-    UIImage *buttonClick = [UIImage imageNamed:@"button_click"];
+    UIImage *defaultClick = [UIImage imageNamed:@"button_click"];
     UIImage *buttonUnclick = [UIImage imageNamed:@"button_unclick"];
     
     // fix '*** -[__NSPlaceholderArray initWithObjects:count:]: attempt to insert nil object from objects[0]'
-    if ((login_btn_normal != nil && login_btn_unable != nil && login_btn_press != nil) || (buttonClick != nil && buttonUnclick != nil)) {
+    if ((login_btn_normal != nil && login_btn_unable != nil && login_btn_press != nil) || (defaultClick != nil && buttonUnclick != nil)) {
       // 登录按钮设置
       model.loginBtnBgImgs = @[
-        login_btn_normal?:[UIImage imageNamed:@"button_click"],
-        login_btn_unable?:[UIImage imageNamed:@"button_unclick"],
-        login_btn_press?:[UIImage imageNamed:@"button_click"]
+        login_btn_normal?:defaultClick,
+        login_btn_unable?:buttonUnclick,
+        login_btn_press?:defaultClick
       ];
     }
   }
@@ -750,10 +772,9 @@
   BOOL checkStatus = [viewConfig boolValueForKey: @"checkBoxHidden" defaultValue: NO];
   model.checkBoxIsHidden = checkStatus;
   if (!checkStatus) {
-    NSArray *checkImagePath = [[viewConfig stringValueForKey: @"checkBoxImages" defaultValue: @","] componentsSeparatedByString:@","];
-    if (checkImagePath.count == 2) {
-      UIImage* unchecked = [self changeUriPathToImage: checkImagePath[0]];
-      UIImage* checked = [self changeUriPathToImage: checkImagePath[1]];
+    UIImage* unchecked = [self changeUriPathToImage: [viewConfig stringValueForKey: @"uncheckedImgPath" defaultValue: nil]];
+    UIImage* checked = [self changeUriPathToImage: [viewConfig stringValueForKey: @"checkedImgPath" defaultValue: nil]];
+    if (unchecked != nil && checked != nil) {
       model.checkBoxImages = @[
         unchecked,
         checked
@@ -768,7 +789,7 @@
   model.changeBtnTitle = [
      [NSAttributedString alloc] initWithString: [viewConfig stringValueForKey: @"changeBtnTitle" defaultValue: @"切换到其他方式"]
      attributes: @{
-       NSForegroundColorAttributeName: [self colorWithHexString: [viewConfig stringValueForKey: @"changeBtnTitleColor" defaultValue: @"#ccc"] alpha: 1],
+       NSForegroundColorAttributeName: [self getColor: [viewConfig stringValueForKey: @"changeBtnTitleColor" defaultValue: @"#ccc"]],
        NSFontAttributeName : [UIFont systemFontOfSize: [viewConfig floatValueForKey: @"changeBtnTitleSize" defaultValue: 18]]
      }
   ];
@@ -1013,6 +1034,26 @@
         }
     }
     return topViewController;
+}
+
++(UIColor *) getColor:(NSString *)hexColor{
+  if (hexColor.length < 8) {
+    return [self colorWithHexString: hexColor alpha: 1];
+  }
+  
+  unsigned int alpha, red, green, blue;
+  NSRange range;
+  range.length =2;
+
+  range.location =1;
+  [[NSScanner scannerWithString:[hexColor substringWithRange:range]]scanHexInt:&alpha];//透明度
+  range.location =3;
+  [[NSScanner scannerWithString:[hexColor substringWithRange:range]]scanHexInt:&red];
+  range.location =5;
+  [[NSScanner scannerWithString:[hexColor substringWithRange:range]]scanHexInt:&green];
+  range.location =7;
+  [[NSScanner scannerWithString:[hexColor substringWithRange:range]]scanHexInt:&blue];
+  return [UIColor colorWithRed:(float)(red/255.0f)green:(float)(green/255.0f)blue:(float)(blue/255.0f)alpha:(float)(alpha/255.0f)];
 }
 
 /**